+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### 1.6.2
+
+#### Features
+
+- add `definition.$type` and `definition.$p0-$p5` helpers for creating type aliases
+
+```typescript
+type DefResult = typeof def.$type; // number
+type DefInput = typeof def.$p0; // string
+
+const def = fn((use, input: string) => 123);
+```
+
 ### 1.6.1
 
 #### Features